+++ lib/src/main/java/ai/freeplay/client/thin/Freeplay.java
@@ -4,6 +4,7 @@ import ai.freeplay.client.HttpConfig;
 import ai.freeplay.client.exceptions.FreeplayConfigurationException;
 import ai.freeplay.client.thin.internal.ThinCallSupport;
 import ai.freeplay.client.thin.resources.feedback.CustomerFeedback;
+import ai.freeplay.client.thin.resources.metadata.Metadata;
 import ai.freeplay.client.thin.resources.prompts.Prompts;
 import ai.freeplay.client.thin.resources.recordings.Recordings;
 import ai.freeplay.client.thin.resources.sessions.Sessions;
@@ -16,6 +17,7 @@ public class Freeplay {
     private final Recordings recordings;
     private final TestRuns testRuns;
     private final CustomerFeedback customerFeedback;
+    private final Metadata metadata;
 
     public Freeplay(FreeplayConfig config) {
         config.validate();
@@ -30,6 +32,7 @@ public class Freeplay {
         recordings = new Recordings(callSupport);
         testRuns = new TestRuns(callSupport);
         customerFeedback = new CustomerFeedback(callSupport);
+        metadata = new Metadata(callSupport);
     }
 
     public Sessions sessions() {
@@ -50,6 +53,10 @@ public class Freeplay {
 
     public CustomerFeedback customerFeedback() {
         return customerFeedback;
+    }
+
+    public Metadata metadata() {
+        return metadata;
     }
 
     public static FreeplayConfig Config() {